[#47]feat: 세션 참여 신청 및 세션 상세 페이지 통합
현재로선 이게 내 세션인지 + 내가 참여를 눌렀던 세션인지를 알 수 없기 때문에, 자신의 세션에 참여 신청으로 인하여 오류가 발생하는 경우 그냥 오류 띄우게만 처리합니다.
