+++ app/(app)/traveler/explore/[id]/_layout.tsx
@@ -1,0 +1,11 @@
+import { Stack } from 'expo-router';
+import React from 'react';
+
+export default function TravelerSessionLayout() {
+  return (
+    <Stack>
+      <Stack.Screen name='index' options={{ headerShown: false }} />
+      {/* <Stack.Screen name='reviews' options={{ headerShown: false }} /> */}
+    </Stack>
+  );
+}
